Move WC sign message request handling out of the VM
